Change equality helper to work with 387 fp unit
When building on i686 without SSE, the ign equality helper uses the extra
precision within the x87 floating point module when comparing the difference
between the two arguments to the epsilon value.  This causes the
Line2Test.CollinearPoint unit test to fail when epsilon is set to .0001

This patch creates an intermediate value for the difference between the two
arguments to be compared, which gets rid of the extra x87 precision. The
resulting comparison is then performed without the extra precision, allowing
the test to pass.

See http://stackoverflow.com/questions/16325277/floating-point-equality-test-and-extra-precision-can-this-code-fail for a discussion of x87 equality pitfalls.