Migrate to relocatable CMake package
Change of the CMake infrastructure to export a relocatable CMake package.

The library is now used in downstream packages as an imported target
(see https://cmake.org/cmake/help/v3.0/manual/cmake-buildsystem.7.html#imported-targets).

This means there is now no need to manually set include directories, link directories and
compiler flags in packages using ignition-math, but it is sufficient to just link
to the imported target, as it is possible to see in the examples directory.